requirements.txt
aiogram>=3.0.0
sqlalchemy[aiosqlite]
python-dotenv
cryptography
apscheduler
